Add status column to EventTable and update permits.blade.php

+++ resources/views/Tableactions/permits.blade.php
@@ -2,14 +2,12 @@
 
     <div class="btn-group mx-2" role="group" aria-label="Third group">
       
-      @if (havePermission(auth()->user(),'event-delete') || $row->user_id == auth()->id())
+      @if ((havePermission(auth()->user(),'event-delete') || $row->user_id == auth()->id()) && $row->status_id != 16)
           <button type="button" class="btn btn-light" wire:click="$dispatch('delete-user', { id: {{$row->id}}})">
             <x-heroicon-o-trash style="width: 28px" />
         </button>
      @endif
 
-     @if (auth()->user()->email == 'eng.ahmed_alghamdi@outlook.com')
-         
-     @endif
+
     </div>
   </div>
